requirements.txt
ja_ginza==3.1.0
seaborn==0.10.1
pandas==1.1.1
plotly==4.9.0
chart_studio==1.1.0
matplotlib==3.3.1
wordcloud==1.8.0
altair_viewer==0.3.0
ginza==3.1.2
spacy==2.3.2
sidetable==0.7.0
ja_ginza_dict==3.1.0
vega==3.4.0
altair==4.1.0
numpy==1.19.1
datapane==0.7.4
Pillow==7.2.0
portpicker==1.3.1
